Moved Grasshopper related extension methods to Grasshopper.Kernel.Extensions as an internal class that is not exposed outside the assembly.
